Refacto API client
