+++ web/src/server-api/Server/Server.js
@@ -2,7 +2,9 @@ import axios from "axios";
 import * as HttpStatus from "http-status-codes";
 import Transform from "./Transform";
 import { Response } from "../Response";
-import { fileFiltersToQueryParams, matchFiltersToQueryParams } from "./helpers";
+import fileFiltersToQueryParams from "./helpers/fileFiltersToQueryParams";
+import clusterFiltersToQueryParams from "./helpers/clusterFiltersToQueryParams";
+import matchesFiltersToQueryParams from "./helpers/matchesFiltersToQueryParams";
 
 export default class Server {
   constructor({ baseURL = "/api/v1", timeout = 10 * 1000, headers = {} } = {}) {
@@ -45,20 +47,13 @@ export default class Server {
     }
   }
 
-  async fetchFileCluster({
-    id,
-    limit = 20,
-    offset = 0,
-    fields = ["meta", "exif", "scenes"],
-    filters,
-  }) {
+  async fetchFileCluster({ id, limit = 20, offset = 0, fields = [], filters }) {
     try {
       const response = await this.axios.get(`/files/${id}/cluster`, {
         params: {
           limit,
           offset,
-          include: fields.join(","),
-          ...matchFiltersToQueryParams(filters),
+          ...clusterFiltersToQueryParams({ filters, fields }),
         },
       });
       const data = this.transform.fetchFileClusterResults(response.data);
@@ -72,14 +67,15 @@ export default class Server {
     id,
     limit = 20,
     offset = 0,
-    fields = ["meta", "exif"],
+    fields = ["meta", "exif", "scenes"],
+    filters = {},
   }) {
     try {
       const response = await this.axios.get(`/files/${id}/matches`, {
         params: {
           limit,
           offset,
-          include: fields.join(","),
+          ...matchesFiltersToQueryParams({ filters, fields }),
         },
       });
       const data = this.transform.fetchFileMatchesResults(response.data);
