+++ web/src/server-api/Server/helpers/matchesFiltersToQueryParams.js
@@ -1,0 +1,10 @@
+/**
+ * Convert file match filters to axios request parameters.
+ */
+export default function matchesFiltersToQueryParams({ fields }) {
+  const params = {};
+  if (fields != null && fields.length > 0) {
+    params.include = fields.join(",");
+  }
+  return params;
+}
